Create layout folder and move layout-related components
- 공통으로 쓰이는 nav, footer을 layout폴더를 생성해서 옮김
- 페이지별로 잘 작동하는지 테스트 완료
