+++ src/components/Layout/Container.module.css
@@ -1,4 +1,4 @@
-@import '../css/reset.css';
+@import '../../css/reset.css';
 
 .container {
   width: 100%;
