+++ src/components/Layout/Footer.module.css
@@ -1,4 +1,4 @@
-@import '../css/reset.css';
+@import '../../css/reset.css';
 
 .footer {
   margin-top: 100px;
