+++ src/components/Layout/Nav.js
@@ -1,9 +1,9 @@
 import { Link, useLocation } from 'react-router-dom';
 import styles from './Nav.module.css';
-import navLogo from '../img/nav_logo.png';
-import navBtn from '../img/nav_btn.png';
-import mobileLoco from '../img/nav_mobile_logo.png';
-import mobileBtn from '../img/nav_mobile_btn.png';
+import navLogo from '../../img/nav_logo.png';
+import navBtn from '../../img/nav_btn.png';
+import mobileLoco from '../../img/nav_mobile_logo.png';
+import mobileBtn from '../../img/nav_mobile_btn.png';
 
 function Nav({ handleResetClick }) {
   const location = useLocation();
