+++ src/components/Layout/Nav.module.css
@@ -1,4 +1,4 @@
-@import '../css/reset.css';
+@import '../../css/reset.css';
 
 .nav {
   position: relative;
